Move useAuth from AuthProvider to Auth
