+++ src/auth/PrivateRoute.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useReducer, useRef } from "react";
 import { RouteProps, Route, useHistory } from "react-router-dom";
 import { Alert, Spinner } from "react-bootstrap";
-import { useAuth } from "./AuthProvider";
+import { useAuth } from "./Auth";
 import Error404 from "../shared/errors/Error404";
 import { APIError } from "../shared/types";
 import ERRORS from "../shared/errors";
